% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_bien.R
\name{praise_bien}
\alias{praise_bien}
\title{Crear frases de éxito estilo mexicanas}
\usage{
praise_bien(exclamaciones = exclamacion_positiva,
  frases = frase_positiva, refranes = refranes_positivos)
}
\arguments{
\item{exclamaciones}{Un vector de caracteres formado por una colección de exclamaciones que expresan felicidad y entusiasmo}

\item{frases}{Un vector de caracteres formado por una colección de frases de celebración}

\item{refranes}{Un vector de caracteres formado por refranes cotidianos de celebración}
}
\value{
Un vector de caracteres
}
\description{
Esta función genera frases chidas a parir de un muestreo aleatorio de exclamaciones y frases, o refranes, comunes en el vocabulario dominguero mexicano.
}
\seealso{
Other crear_frases: \code{\link{praise_error}}
}
\concept{crear_frases}
