% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_crear_emi.R
\name{praise_crear_emi}
\alias{praise_crear_emi}
\title{praise_crear_emi}
\usage{
praise_crear_emi(exclamaciones, frases, refranes, comidas = comida_rica)
}
\arguments{
\item{exclamaciones}{Un vector de caracteres formado por una colección de exclamaciones que expresan felicidad y entusiasmo}

\item{frases}{}

\item{refranes}{}

\item{comidas}{}
}
\description{
praise_crear_emi
}
