% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_error.R
\name{praise_error}
\alias{praise_error}
\title{Crear frases de error estilo mexicanas}
\usage{
praise_error(exclamaciones = exclamacion_negativa,
  frases = frase_negativa, refranes = refranes_negativos)
}
\arguments{
\item{exclamaciones}{Un vector de caracteres formado por una colecci贸n de exclamaciones que expresan frustraci贸n}

\item{frases}{Un vector de caracteres formado por una colecci贸n de frases para motivar un poco}

\item{refranes}{Un vector de caracteres formado por refranes cotidianos de fracaso}
}
\value{
Un vector de caracteres
}
\description{
Esta funci贸n genera frases chidas a parir de un muestreo aleatorio de exclamaciones y frases, o refranes, comunes en el vocabulario dominguero mexicano.
}
\seealso{
Other crear_frases: \code{\link{praise_bien}}
}
\concept{crear_frases}
