% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_praisemx.R
\name{set_praisemx}
\alias{set_praisemx}
\title{Modificar las opciones del ambiente de R
para imprimir un mensaje de error adicional
con español mexicano}
\usage{
set_praisemx()
}
\description{
Modificar las opciones del ambiente de R
para imprimir un mensaje de error adicional
con español mexicano
}
