% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_crear_emi.R
\name{praise_crear_emi}
\alias{praise_crear_emi}
\title{praise_crear_emi}
\usage{
praise_crear_emi(exclamaciones, frases, refranes, comidas = comida_rica)
}
\arguments{
\item{exclamaciones}{Es un vector de caracteres formado por una colecciones de exclamaciones populares mexicanas}

\item{frases}{Es un vector de caracteres formado por una colecciones de frases que suelen expresar los mexicanos en ciertas situaciones.}

\item{refranes}{Es un vector de caracteres formado por una colecciones de refranes que los mexicanos transmiten verbalmente a traves de las generaciones}

\item{comidas}{Es un vector de caracteres que contiene una colección de comida típica mexicana}
}
\value{
Una oración construida por refranes y exclamaciones del vocabulario mexicano
}
\description{
Esta función genera frases de motivación en función de un error o un éxito al corres un comando. Estas frases provienen del vocabulario dominguero mexicano.
}
