% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_error.R
\name{praise_error}
\alias{praise_error}
\title{Crear frases de error estilo mexicanas}
\usage{
praise_error(exclamaciones = exclamacion_negativa,
  frases = frase_negativa, refranes = refranes_negativos)
}
\arguments{
\item{exclamaciones}{Colección de exclamaciones que expresan fracaso}

\item{frases}{Colección de frases para motivar un poco}

\item{refranes}{Refranes cotidianos de fracaso}
}
\description{
Esta función genera diferentes frases a parir de un muestreo aleatorio de exclamaciones y frases, o refranes, comunes en el vocabulario dominguero mexicano.
}
